[TASK] Use universal link list to render all link lists
