--- resources/js/Components/LinkList/EmptyState.svelte
@@ -1,10 +1,0 @@
-<li class="flex items-center justify-center text-center h-full">
-    <div>
-        <svg class="mx-auto h-12 w-12 text-gray-400" xmlns="http://www.w3.org/2000/svg"
-             fill="none" viewBox="0 0 24 24" stroke="currentColor" stroke-width="1.5">
-            <path stroke-linecap="round" stroke-linejoin="round"
-                  d="M13.828 10.172a4 4 0 00-5.656 0l-4 4a4 4 0 105.656 5.656l1.102-1.101m-.758-4.899a4 4 0 005.656 0l4-4a4 4 0 00-5.656-5.656l-1.1 1.1"/>
-        </svg>
-        <h3 class="mt-2 text-sm font-medium text-gray-900">No links</h3>
-    </div>
-</li>
